+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Hooked up the `Download plan` page and added a `download-narrative` api endpoint [#313]
 
 ### Updated
+- Made improvements to auth handling in `middleware` and `authHelper.ts` [#1035]
 - Made text changes to upcoming blog posts [#989]
 - For Research Outputs, updated repositories and metadata standards fields to be automatically enabled when user selects custom ones [#943]
 - Updated Licenses and Output Types to use label "Use custom list" [#943]